don't trigger on internal temp-names
