+++ lib/fs/basicfs_test.go
@@ -12,7 +12,6 @@ import (
 	"path/filepath"
 	"runtime"
 	"sort"
-	"strings"
 	"testing"
 	"time"
 )
@@ -305,38 +304,6 @@ func TestUsage(t *testing.T) {
 	}
 }
 
-func TestWindowsPaths(t *testing.T) {
-	if runtime.GOOS != "windows" {
-		t.Skip("Not useful on non-Windows")
-		return
-	}
-
-	testCases := []struct {
-		input        string
-		expectedRoot string
-		expectedURI  string
-	}{
-		{`e:\`, `\\?\e:\`, `e:\`},
-		{`\\?\e:\`, `\\?\e:\`, `e:\`},
-		{`\\192.0.2.22\network\share`, `\\192.0.2.22\network\share`, `\\192.0.2.22\network\share`},
-	}
-
-	for _, testCase := range testCases {
-		fs := newBasicFilesystem(testCase.input)
-		if fs.root != testCase.expectedRoot {
-			t.Errorf("root %q != %q", fs.root, testCase.expectedRoot)
-		}
-		if fs.URI() != testCase.expectedURI {
-			t.Errorf("uri %q != %q", fs.URI(), testCase.expectedURI)
-		}
-	}
-
-	fs := newBasicFilesystem(`relative\path`)
-	if fs.root == `relative\path` || !strings.HasPrefix(fs.root, "\\\\?\\") {
-		t.Errorf("%q == %q, expected absolutification", fs.root, `relative\path`)
-	}
-}
-
 func TestRooted(t *testing.T) {
 	type testcase struct {
 		root   string
@@ -480,36 +447,5 @@ func TestRooted(t *testing.T) {
 			t.Errorf("Unexpected pass for rooted(%q, %q) => %q", tc.root, tc.rel, res)
 			continue
 		}
-	}
-}
-
-func TestWindows83(t *testing.T) {
-	if runtime.GOOS != "windows" {
-		t.Skip("only relevant on windows")
-	}
-
-	fs, dir := setup(t)
-	defer os.RemoveAll(dir)
-
-	shortAbs, _ := fs.rooted("LFDATA~1")
-	long := "LFDataTool"
-	longAbs, _ := fs.rooted(long)
-	deleted, _ := fs.rooted(filepath.Join("foo", "LFDATA~1"))
-	notShort, _ := fs.rooted(filepath.Join("foo", "bar", "baz"))
-
-	fd, err := fs.Create(long)
-	if err != nil {
-		t.Fatal(err)
-	}
-	fd.Close()
-
-	if res := fs.resolveWin83(shortAbs); res != longAbs {
-		t.Errorf(`Resolving for 8.3 names of "%v" resulted in "%v", expected "%v"`, shortAbs, res, longAbs)
-	}
-	if res := fs.resolveWin83(deleted); res != filepath.Dir(deleted) {
-		t.Errorf(`Resolving for 8.3 names of "%v" resulted in "%v", expected "%v"`, deleted, res, filepath.Dir(deleted))
-	}
-	if res := fs.resolveWin83(notShort); res != notShort {
-		t.Errorf(`Resolving for 8.3 names of "%v" resulted in "%v", expected "%v"`, notShort, res, notShort)
 	}
 }
